requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.26.82
botocore==1.29.82
celery==5.2.7
certifi==2022.6.15
cffi>=1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1
django-allauth==0.51.0
django-celery-beat==2.4.0
django-cleanup==7.0.0
django-cors-headers==3.13.0
django-environ==0.10.0
django-filter==22.1
django-simple-history==3.2.0
django-storages==1.13.2
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
greenlet>=2.0.2
gunicorn==20.1.0
idna==3.3
jmespath==1.0.1
kombu==5.2.4
oauthlib==3.2.2
pillow>=8.0.0

prompt-toolkit==3.0.36
psycopg2-binary

pycparser==2.21
PyJWT==2.6.0
pyotp==2.8.0
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
redis==4.5.1
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==2.0.3
sqlparse==0.4.2
typing_extensions==4.4.0
tzdata==2022.7
Unidecode==1.3.5
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
